+++ shuup/front/apps/auth/urls.py
@@ -7,6 +7,8 @@
 # LICENSE file in the root directory of this source tree.
 from django.conf.urls import patterns, url
 
+from shuup.core.utils.maintenance import maintenance_mode_exempt
+
 from .views import (
     LoginView, LogoutView, RecoverPasswordCompleteView,
     RecoverPasswordConfirmView, RecoverPasswordSentView, RecoverPasswordView
@@ -24,12 +26,12 @@ urlpatterns = patterns(
         RecoverPasswordView.as_view(),
         name='recover_password'),
     url(r'^recover-password/(?P<uidb64>.+)/(?P<token>.+)/$',
-        RecoverPasswordConfirmView.as_view(),
+        maintenance_mode_exempt(RecoverPasswordConfirmView.as_view()),
         name='recover_password_confirm'),
     url(r'^recover-password/sent/$',
         RecoverPasswordSentView.as_view(),
         name='recover_password_sent'),
     url(r'^recover-password/complete/$',
-        RecoverPasswordCompleteView.as_view(),
+        maintenance_mode_exempt(RecoverPasswordCompleteView.as_view()),
         name='recover_password_complete'),
 )